+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/interceptor/analyse/BehaviorAnalyseAdvisor.java
@@ -55,6 +55,4 @@ public class BehaviorAnalyseAdvisor {
             LOGGER.debug("方法 " + className + "." + methodName + " 运行结束，用时 " + (lastTimeStamp - firstTimeStamp) + " 毫秒");
         }
     }
-
-
 }